+++ src/m_config.c
@@ -157,7 +157,6 @@ static default_t cvars[] =
     CONFIG_VARIABLE_INT          (r_liquid_lowerview,                                BOOLVALUEALIAS    ),
     CONFIG_VARIABLE_INT          (r_liquid_swirl,                                    BOOLVALUEALIAS    ),
     CONFIG_VARIABLE_OTHER        (r_lowpixelsize,                                    NOVALUEALIAS      ),
-    CONFIG_VARIABLE_INT          (r_messagescale,                                    SCALEVALUEALIAS   ),
     CONFIG_VARIABLE_INT          (r_mirroredweapons,                                 BOOLVALUEALIAS    ),
     CONFIG_VARIABLE_INT          (r_playersprites,                                   BOOLVALUEALIAS    ),
     CONFIG_VARIABLE_INT          (r_rockettrails,                                    BOOLVALUEALIAS    ),
@@ -257,8 +256,7 @@ valuealias_t valuealiases[] =
     { "red",       1, BLOODVALUEALIAS     }, { "all",     2, BLOODVALUEALIAS     },
     { "imperial",  0, UNITSVALUEALIAS     }, { "metric",  1, UNITSVALUEALIAS     },
     { "off",       0, CAPVALUEALIAS       }, { "none",   -1, SKYVALUEALIAS       },
-    { "off",      -1, SKYVALUEALIAS       }, { "small",   0, SCALEVALUEALIAS     },
-    { "big",       1, SCALEVALUEALIAS     }, { "none",    5, FACEBACKVALUEALIAS  },
+    { "off",      -1, SKYVALUEALIAS       }, { "none",    5, FACEBACKVALUEALIAS  },
     { "off",       5, FACEBACKVALUEALIAS  }, { "none",    0, ARMORTYPEVALUEALIAS },
     { "green",     1, ARMORTYPEVALUEALIAS }, { "blue",    2, ARMORTYPEVALUEALIAS },
     { "",          0, NOVALUEALIAS        }
@@ -752,9 +750,6 @@ static void M_CheckCVARs(void)
 
     if (r_liquid_swirl != false && r_liquid_swirl != true)
         r_liquid_swirl = r_liquid_swirl_default;
-
-    if (r_messagescale != r_messagescale_small && r_messagescale != r_messagescale_big)
-        r_messagescale = r_mirroredweapons_default;
 
     if (r_mirroredweapons != false && r_mirroredweapons != true)
         r_mirroredweapons = r_mirroredweapons_default;